+++ lib/page/mytests/context_study/myapp_context_study.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:logger/logger.dart';
-import 'package:riverpod_modifiers_example/page/mytests/context_study/lixopage_context_sutdy.dart';
-import 'package:riverpod_modifiers_example/page/mytests/context_study/utils.dart';
+import 'package:riverpod_modifiers_example/page/mytests/context_study/utils/utils.dart';
 import 'package:riverpod_modifiers_example/page/mytests/theme_study/extension/theme_provider.dart';
 
-import 'homepage_theme_study.dart';
-import 'lixopage2_context_sutdy.dart';
+import 'pages/homepage_theme_study.dart';
+import 'pages/lixopage2_context_sutdy.dart';
+import 'pages/lixopage_context_sutdy.dart';
 
 final myLogger = Logger();
